extract compose/yaml/dictionary to seperate file
