+++ providers/docker/compose/dictionary_test.go
